+++ daemon-tests/Cargo.toml
@@ -16,6 +16,7 @@ rust_decimal = "1.20"
 rust_decimal_macros = "1.20"
 time = "0.3"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
+tokio-tasks = { path = "../tokio-tasks" }
 tracing = { version = "0.1" }
 tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt", "ansi", "env-filter", "local-time", "tracing-log", "json"] }
 xtra = { version = "0.6" }
