+++ daemon-tests/src/lib.rs
@@ -27,7 +27,6 @@ use daemon::projection::Feeds;
 use daemon::seed::RandomSeed;
 use daemon::seed::Seed;
 use daemon::MakerActorSystem;
-use daemon::Tasks;
 use daemon::HEARTBEAT_INTERVAL;
 use daemon::N_PAYOUTS;
 use daemon::SETTLEMENT_INTERVAL;
@@ -39,6 +38,7 @@ use std::str::FromStr;
 use std::time::Duration;
 use tokio::net::TcpListener;
 use tokio::sync::watch;
+use tokio_tasks::Tasks;
 use tracing::subscriber::DefaultGuard;
 use tracing_subscriber::filter::LevelFilter;
 use tracing_subscriber::util::SubscriberInitExt;
