+++ daemon/Cargo.toml
@@ -36,6 +36,7 @@ sqlx = { version = "0.5", features = ["offline", "sqlite", "uuid", "runtime-toki
 thiserror = "1"
 time = { version = "0.3", features = ["serde", "macros", "parsing", "formatting"] }
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
+tokio-tasks = { path = "../tokio-tasks" }
 tokio-tungstenite = { version = "0.15", features = ["rustls-tls"] }
 tokio-util = { version = "0.6", features = ["codec"] }
 tracing = { version = "0.1" }
