+++ maker/Cargo.toml
@@ -18,6 +18,7 @@ rust-embed-rocket = { path = "../rust-embed-rocket" }
 serde = { version = "1", features = ["derive"] }
 shared-bin = { path = "../shared-bin" }
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
+tokio-tasks = { path = "../tokio-tasks" }
 tracing = { version = "0.1" }
 uuid = "0.8"
 xtra = { version = "0.6" }
