+++ taker/Cargo.toml
@@ -20,6 +20,7 @@ serde = { version = "1", features = ["derive"] }
 shared-bin = { path = "../shared-bin" }
 time = "0.3"
 tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "net"] }
+tokio-tasks = { path = "../tokio-tasks" }
 tracing = { version = "0.1" }
 uuid = "0.8"
 x25519-dalek = "1.1"
