+++ tokio-tasks/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "tokio-tasks"
+version = "0.1.0"
+edition = "2021"
+description = "Control the lifetime of async tasks by leveraging Drop"
+
+[dependencies]
+futures = { version = "0.3", default-features = false, features = ["std"] }
+tokio = { version = "1", default-features = false, features = ["rt-multi-thread"] }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["time", "macros"] }
